[SYCL][HIP] Add managed memory USM checks for HIP prefetch (#10761)
Adds `hipDeviceAttributeManagedMemory` attribute check to verify the HIP
device supports managed memory in order to use the prefetch API as well
as checking the pointer to migrate is actually a managed allocation
(USM/SVM).

If the check fails we return early and set a warning message (with
`UR_RESULT_SUCCESS` as return code).

As a follow-up. we may need to detect system-allocated memory and either
continue with just a warning message or throw a more targeted error
message and exit.