Fix TopologicalSort for Joint and Cooperative matrices
visit method of the sort relies on getNonLiteralOperands method of the
SPIRVType which is being inserted in the module. Without it dependent
types can be inserted in the module in incorrect order.

For example:
TypeCooperativeMatrixKHR %ID%
TypeStruct ... %ID%
is the correct order, but without the patch in some cases the translator
could generate the opposite order.

It's a cherry-pick of
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/pull/2143

Signed-off-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>