[SYCL][Bindless] Fix for fill_rand
The `fill_rand` helper function calls `uniform_int_distribution`
with different data types that are being tested.
The C++ standard doesn't allow passing `char`
or any similar types to it - this currently work with GCC,
but (correctly) fails with MSVC.

This patch changes the type used for the RNG
depending on the tested data type.