[SYCL][FPGA] Allow max_interleaving with disable_loop_pipelining attribute on same loop
The front-end currently issues mutually exclusive error when [[intel::max_interleaving()]]
and [[intel::disable_loop_pipelining]] both attributes apply on the same loop.

We should not error out in this situation - the attributes are NOT mutually exclusive.
Our middle end and backend already support having both features enabled and/or disabled in
any combination.

This patch removes the restriction.

Signed-off-by: Soumi Manna <soumi.manna@intel.com>