[SYCL][NFC] Make UR_CHECK_ERROR a void return function
`UR_CHECK_ERROR` was designed to return `ur_result_t`, however in
practice it was guaranteed to only ever return `UR_RESULT_SUCCESS`, as
other paths would either terminate, abort or throw.

This in turns leads to poor quality/error prone code, as the codebase
was littered with:
* statements not checking the return value - depending on the compiler
  generating a warning,
* extra check on the return which was only ever going to be true.

Some care was required, as the codebase has a habit of accumulating err
codes across branches, so depending on the use case the initial value of
`ur_result_t Result`s had to be set accordingly (now that
`UR_CHECK_ERROR` does not return).