[Driver][SYCL] Enable early AOT abilities when creating objects
Adds support for enabling the ability to perform a full device
compilation and link.  Use the -ftarget-device-link option when also
compiling to object with -c.  This will trigger the additional device
linking steps to be performed after the device compilation is completed.

Currently only supported for spir64_gen targets, as we rely on the
addition of the -device <arg> values to differentiate the fat objects
that are generated.

Upon consumption of these new fat objects, the driver will scan for
these unique binaries and instead of going through the device link,
these binaries will be sent directly to the host link step to be added
to the final executable.

Performs some refactoring of the device link code, allowing for a
common platform for compile and link to access.