[SYCL][UR][L0] Add support for multi-device kernel compilation
urProgramBuild and alike dont accept device as parameters, so
it is expected the programs and kernels are built for all devices
in the context.

With this patch, the L0 modules and L0 kernels are created
for all L0 devices in the context, and then on enqueue operations,
the correct kernel is used matching the device in the queue.

Signed-off-by: Jaime Arteaga <jaime.a.arteaga.molina@intel.com>