[SYCL][Doc] Propose renaming device_global
Rename device_global to device_specific, to avoid confusion between "at global
scope" and "in the global address space", and to align with the naming of
work_group_specific.

This commit only renames the class. Additional work may be required to further
align the concepts of device-specific and work-group-specific variables (e.g.
by permitting device-specific variables to be declared at function scope).

Signed-off-by: John Pennycook <john.pennycook@intel.com>