[SYCL][ESIMD] Add hidden compiler option to not compile host code (#11037)
ESIMD code can only be run on an Intel GPU. Previously we had the ESMID
emulator that called into host implementations of some of the GPU
intrinsics, but that has been deprecated and will be removed. Even when
the emulator was supported, it was seldom used by customers.

To summarize, the host compile for ESIMD code had little value to
customers previous, and will soon have zero value once the emulator is
removed.

However, the host compile can be slow. We have a bug report from a
customer reporting slow compilation time for an ESIMD kernel with a
large vector. The reason for the slowdown was investigated by the X86
backend team and they found for the host compile, the vector exceeds the
maximum legal vector size on X86 and instruction selection needs to
spend a lot of time doing heavy lifting to generate the right assembly.

This is a total waste of the user's time as that code will never and can
never be run.

As a workaround for these cases, add a new hidden option,
`-fsycl-esimd-build-host-code`/`-fno-sycl-esimd-build-host-code`,
default `-fsycl-esimd-build-host-code`, to try to speed up the host
compile. This is mostly done through a new pass which runs early and
removes all IR for ESIMD functions and replaces them with a single
return statement. This allows later optimizations to do even more
cleanup.

Also, I modified some of the ESIMD macros to generate IR that is more
likely to be optimized out for the host, just disabling inlining and
setting internal linkage.

We do not intend to advertise this option as there should be no issue in
most cases, but we would like to be able to provide a workaround if
users hit an issue.

I have manually verified this option fixes the slowdown reported by the
customer.

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>