[SYCL][KernelFusion] Ensure queue_impl is not unnecessarily referenced (#10948)
The KernelFusionCmd keeps a references to the queue on which it was
submitted. However these commands are also kept on the side in order to
ensure event validity and are only reset when start fusion is called
again on the same queue.
This forces queue_impl and events to be kept alive as the command keeps
the last reference until the sycl shutdown procedure.

The patch forces 1) the command to reset it pointers to the queue once
enqueued and 2) to clean up the KernelFusionCmd attached to the queue to
free up events.

Note: there is no new test as kernel fusion already cover this and I'm
not too sure on how to approach it. I'm happy to try out any suggestion.

---------

Signed-off-by: Victor Lomuller <victor@codeplay.com>