[SYCL] Fix propagation of aspects from global variables (#11173)
Before this PR, the `pf-wg-atomic64.cpp` (from this PR) fails because
the `atomic64` aspect was not propagated to the calling kernel. The
reason is:

- With opaque pointers disabled, since the `this` parameter of the
constructor of `atomic_ref` would have a type of `atomic_ref*`, the call
to the constructor of `atomic_ref` would be enough to propagate the
`atomic64` aspect to the kernel. With opaque pointers enabled, this
would not be the case: the `this` parameter would just be a `ptr`.
- Additionally, for some reason I am unsure of, the code generated for
`parallel_for_work_group` (and not, say, for `parallel_for`) has the
pointer for `feature` generated as a global variable instead of
generated from an `alloca` instruction. If there were an `alloca` of an
`atomic_ref`, there would be no problem.

Now, when examining the operands of instructions in
`SYCLPropagateAspectsPass`, we now consider if the operand is a global
variable, and if so, propagate the type is points to instead of
propagating the pointer type.