[SYCL] Don't include <cmath> from <sycl/sycl.hpp>
This follows #11196 that did the same with <complex> before. We still want to
support `std::` math functions in user code (as part of
`sycl/doc/extensions/supported/C-CXX-StandardLibrary.rst`). Also, often times
STL implements math functions support somewhat like this:

```
extern "C" {
 extern double cos (double __x) noexcept (true);
}
extern "C++"
{
namespace std __attribute__ ((__visibility__ ("default")))
{
 using ::cos;
}
}
```

As such, we still want to let the compiler know which symbols in the global
namespace are known (e.g. `::cos`) and which aren't and need an error ( `SYCL
kernel cannot call an undefined function without SYCL_EXTERNAL attribute`).

One option to implement that is to recognize those functions as builtins in the
front end (which #11014 tries to do). Another approach, taken in this PR, is to
keep providing declaration in SYCL headers but only do so when customer included
`<cmath>` explicitly (via `#include_next`, similarly to `<complex>` support
added in #11196).