[SYCL] Add esimd as an optional device feature (#10017)
This change adds a new aspect for esimd, `ext_intel_esimd`, and
annotates the two fundamental esimd classes, `simd_obj_impl` and
`simd_view_impl` with the `uses_aspect` attribute.

`simd_obj_impl` is the base class of `simd` and `simd_mask` which are
the fundamental user-facing classes.
`simd_view_impl` is the base of only `simd_view`. `simd_obj_impl` is not
a base of `simd_view_impl`, but every `simd_view_impl` requires a
reference to a `simd` or `simd_mask` at construction time, so I am not
sure if we truly need to annotate `simd_view_impl`, but I added it to be
safe.

It also adds a new PI device info query,
`PI_EXT_INTEL_DEVICE_INFO_ESIMD_SUPPORT` that is used to query at
runtime if a device supports ESIMD. For UR-based plugins, we map that to
`UR_DEVICE_INFO_ESIMD_SUPPOR`.

The implementation simply returns false for cuda, hip and native_cpu.
For l0 and opencl, we check that the device is an intel gpu by querying
the device type is gpu and the vendor id is `0x8086`.
For ESIMD emulator we simply return true.

I would appreciate careful review on the plugin changes in particular,
as I am not an expert.

This change also updates the esimd spec to document the new aspect.

In a future change, I plan to use the new aspect to remove the
requirement for the `SYCL_ESIMD_FUNCTION` and `SYCL_ESIMD_KERNEL` macros
that set function attributes, but I am not doing that as part of this PR
as it requires more investigation.

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>