[do not review][Driver] Adjust integration footer inclusion
This is an experimental change right now - just testing.

Remove usage of the 'append-file' tool which creates a temporary file
that is used when building the host source file in a SYCL offloading
compilation environment.

Instead, pull in the main source file as an '-include' and the footer is
then used as the main source file.