[vecz] Fix alignment of load operations
We were alternatively under- and over-setting the alignment on
vectorized loads. Under-setting the alignment is still correct if
sub-optimal.

However, the `SimplifyMaskedMemOpsPass` was ignoring the alignment of
the scalar masked mem load/store builtins when optimizing to regular
LLVM load and store instructions. This would over-set the alignment,
meaning the compiler could assume properties about the address which
were not true.

This led to riscv reporting a misaligned store exception on some SYCL
CTS tests.