[CI] Lock Windows CI (#11240)
I've created `WindowsCILock` environment and put
@intel/dpcpp-devops-reviewers into *workflow* reviewers list. Tasks
using this environment will need to be approved by a member of this team
before they run.

Once CI is stabilized I'm going to clear the box to require those
reviews but keep the workflow files change so that we can lock the CI in
future. I'm also planning in implementing the same for Linux pre-commit
later.