[SYCL][HIP] Fix hip prefetch/mem_advise (noop) for rocm < 5.0 (#10958)
The main fix here is to enable prefetch functionality when the HIP
backend is built with rocm versions < 5.
The change to the prefix/mem_advise UR hip adapter APIs fixes four
e2e-tests that fail test-e2e only for HIP 4.x versions (These failures
don't come up in the CI since it tests using rocm 5.x).

The change to Tracing/image_printers.cpp is effectively re-XFailing this
test for rocm 4 only. rocm 4 doesn't support `hipCreateSurfaceObject`
which is called by this test. I'm not sure this legacy image
functionality (Which as I understand it will be replaced by bindless
images) is really working for rocm 5 in a meaningful way, since most
other legacy image e2e tests are XFAIL for hip. But this test can still
be useful for us for ROCM 5 compatibility testing. The XFAIL was
recently removed here
https://github.com/intel/llvm/commit/745febe51b68b2a7cc0ed021e6ebb863af48c4fa
which led to us finding the rocm 4 fail. I guess that the CI at some
point switched from testing rocm 4 to testing rocm 5, which meant this
test stopped failing in CI.

---------

Signed-off-by: Jack Kirk <jack.kirk@codeplay.com>