[vecz] Fix invalid dangling AssumptionCaches
The StrideAnalysisResult had its own local copy of an AssumptionCache.

These were still live at the time we deleted instructions at the end of
packetization, and/or tore down the vectorization process upon
packetization failure.

This meant that some llvm Values could have callback handles in two
assumption caches, and when one was deleted the other would hold onto
garbage memory.

The fix is to use the same assumption cache in StrideAnalysisResult as
is obtained from the FunctionAnalysisManager.

An additional bug was that we were ostensibly invalidating all function
analyses on deleted vectorized functions, but were accidentally
inverting the set of preserved analyses such that all analyses were
being *preserved*, rather than invalidated.