Remove `id` from list of disallowed uniform types
We decided that `sycl::id` should not be a disallowed uniform type.
It not allowed to pass a work-item's ID to `uniform` because that
value does not have a uniform value for all work-items.  However,
the application can construct its own `sycl::id` in a uniform way,
so we should not disallow `sycl::id` in general.

This commit restores the spec wording and also changes the
implementation to align with that wording.