[test] improve match file error output
When a test fails a match, cmakes print both the
input file and the match file, so that it's easy to see
why the failure happened and how to fix it. Unfortunately,
cmake was inserting 3 newlines between every line of both
input and match files, making things hard to read.

This patch fixes this by moving match and input files printing
to the python script.

Co-authored-by: Petr Vesely <petr.vesely@codeplay.com>