[SYCL] Added `ext_oneapi_non_uniform_groups` aspect (#10902)
The errors that you get if you try to e.g. create a `ballot_group` in
e.g. HIP backend are not instructive. This PR introduces
`ext_oneapi_non_uniform_groups` and uses the
`[[__sycl_detail__::__uses_aspects__(aspect::foo)]]` annotations
detailed in:


https://github.com/intel/llvm/blob/sycl/sycl/doc/design/OptionalDeviceFeatures.md#changes-to-dpc-headers

to give a clear runtime error in the case that a programmer tries to
execute a kernel that uses non-uniform groups on a device that does not
supported the non-uniform group extension.

---------

Signed-off-by: Jack Kirk <jack.kirk@codeplay.com>
Co-authored-by: Greg Lueck <gregory.m.lueck@intel.com>