[clang-linker-wrapper][SYCL Offload] Modify SYCL offloading flow to use clang-linker-wrapper (#11187)
This change is driven by the following design document:
sycl/doc/design/OffloadDesign.md

All actions that include (1) linking of all device offload regions for
SYCL offload kind, (2) running the post-link tool, (3) translating the
LLVM IR to SPIRV and (4) wrapping the device code inside a host object
have been moved from the driver to the clang-linker-wrapper tool.

This PR aims to establish a baseline implementation to enable the 'new
offloading' model for SYCL offload. Changes in this PR will enable us to
use clang-linker-wrapper tool for compiling a SYCL offload fat binary.
Driver changes are required to call clang-linker-wrapper with
appropriate options and will be added in subsequent PRs.

A simple test has been provided to check for the correct compilation
flow. More testing will be added as the design takes shape.

This PR is the first step in many steps required to modify SYCL
offloading to use clang-linker-wrapper and should be considered a
work-in-progress when reviewing. All changes are under an existing
option that is turned off by default (--offload-new-driver) and hence
will not be invoked by default.

Thanks in advance for feedback.
Sincerely

---------

Signed-off-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>