[SYCL][Doc] Move uniform extension spec (#11259)
The uniform extension was implemented in #5871.  Move the extension
specification document to the "experimental" directory, and update to
the
latest specification template.

Also fix a small bug in the implementation.  The `sycl::id` type should
not be disallowed for use in `uniform`.  Although, it is invalid to
pass the work-item's ID to `uniform` (because that value is not uniform
across all work-items), the user could construct their own `id` that
does have a uniform value.