[SYCL][ESIMD][E2E] Fix noinline_args tests after pulldown
After the recent pulldown, clang generates llvm.lifetime calls for
the `add` function arguments in different locations causing us to hit a known optimizable
case in ESIMDOptimizeVecArgCallConv so we start passing the arguments
by value. This ends up causing different stack size, and these
tests are very sensitive to stack size.

Add a noop inline assembly statement to prevent the optimization from
kicking in so we preserve by-pointer argument passing in this case.

Closes: https://github.com/intel/llvm/issues/11036

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>