[ur] Document the Adapter Change Process
Now that adapter implementations are coming to live in this repo we need
a process to ensure we maintain the quality of changes in dependent
projects such as [intel/llvm](https://github.com/intel/llvm). This patch
updates the contribution guide to include a section titled *Adapter
Change Process* which details the process to follow when making changes
which affect the API/ABI or UR or the behaviour of adapters.