[SYCL][CI] Run check-sycl-unittests on Linux CI
OsUtils.cpp is removed because it's not reliable. We build the test
using "RUNPATH" and not "RPATH", so LD_LIBRARY_PATH has precedence over
it. However, in post-commit task we use a nightly intel/llvm build set
through LD_LIBRARY_PATH. That causes the unit-test's executable to load
/opt/sycl/'s libsycl.so and not the one the test expects.

It might be possible to properly fix this issue, but I don't know how
and, IMO, it's more beneficial to remove one unit test and enable the
rest than not run anything at all.