[SYCL][FPGA] Rename [[intel::disable_loop_pipelining]] attribute metadata
Metadata emitted from front-end when an FPGA IPA kernel has the property pipelined<N> is a function metadata of the form !disable_loop_pipelining !X, !X = !{i32 0|1}.

This is confusing as the user means to pipeline the kernel/not pipeline the kernel as opposed to enabling/disabling the pipelining of all loops in the kernel/function, and this name does not reflect that. This patch renames the functin metadata to !disable_kernel_pipelining.

This was likely caused by a miscommunication with the related loop metadata that's currently being attached when the loop attribute [[intel::disable_loop_pipelining]] is used: {!"llvm.loop.intel.pipelining.enable", i32 0}

This patch also changes the loop metadata to !llvm.loop.intel.pipelining.disable and invert the argument to reflet the correct attribute behavior and match with function metadata change.

Signed-off-by: Soumi Manna <soumi.manna@intel.com>