[SYCL] Selection of device images where spec consts are replaced with default
Now SYCL tools can generate device image where specialization constants are
replaced with default values in addition to original device image.
This patch implements selection of such device image when possible. It
is possible when specialization constant values are not set for the
device image.
When specialization constants are used we always work with kernel_bundle
and obtain kernel object to enqueue via get_kernel method of
the kernel_bundle. This is the point when device image is selected.
Before we were selecting the first image containing the kernel. Now
image selection was modified to take into account new device images
where specialization constants are replaced with default values.