Describe interaction with type-name queries
Note that the forms of `has_kernel` and `get_kernel` that take a kernel
type-name may be used to query for a kernel defined in the kernel
compiler.  However, kernels in OpenCL C cannot use these features.

This is added for extensibility for when we allow kernels to be written
in SYCL.