[SYCL] Fix test/check_device_cod/fpga_*_local.cpp for asserts disabled in compiler (#11423)
Caused by https://github.com/intel/llvm/pull/10849 

Due to the fact that if the test are run on a compiler that was compiled
with `--hip --cuda`. This causes the LLVM IR variable names to be
obfuscated from `%empty` to `%0`. The failing LIT tests relied on that
names being intact.