Add AMDGPU reflect pass for atomic xor
AMDGPU reflect pass is needed to choose between safe and unsafe atomics
at the libclc level. In the long run we will delete this patch as work
is being done to ensure correct lowering of atomic instructions. See
patches:

https://github.com/llvm/llvm-project/pull/85052/
https://github.com/llvm/llvm-project/pull/69229/

This work is necessary as malloc shared atomics rely on PCIe atomics
which can have patchy and unreliable support. We want to therefore be
able to choose at compile time whether we should use safe atomics using
CAS (which PCIe should support), or if we want to rely of the
availability of the newest PCIe atomics, if malloc shared atomics are
desired.

Also changes the implementation of Or, And so that they can choose
between the safe or unsafe version based on the AMDGPU reflect value.