[SYCL][LIBCLC] Remove target features attributes when compiling for AMD (#11247)
AMDGPU remove incompatible functions pass replaces all uses of functions
that use GPU features incompatible with the current GPU with null then
deletes the function. This didn't use to cause problems, as all of
libclc functions were inlined prior to incompatible functions pass. Now
that the inliner runs later in the pipeline we have to remove all of the
target features, so libclc functions will not be earmarked for deletion.

NOTE: I've also reverted the pulldown revert commit to make sure that
the original error condition is recreated, hence this PR should not be
squashed prior to merging.

Fixes: https://github.com/intel/llvm/issues/10816