[SYCL][Fusion] Support heterogeneous ND ranges on CUDA (#11421)
This PR splits up the builtin remapping approach into a
target-independent part (remains in `Remapper` class) and
target-specific hooks (added to `TargetFusionInfo` class). The existing
SPIR-V implementation is moved to these hooks, and we add a new
implementation for the PTX backend.

The main challenge is the remapping of global sizes and IDs, as the PTX
backend does not have corresponding intrinsics and we only see the
lowered computations. We solve this by either honoring the local size
provided during an nd_range launch, or simulating one "large" workgroup
with the same size as the original global size otherwise.

---------

Signed-off-by: Julian Oppermann <julian.oppermann@codeplay.com>