[ESIMD] Fix performance regression caused by using LLVM IR for block load/store
The regression is seen when block_load/block_store() and copy_to()/copy_from()
are used in ESIMD functions called through invoke_simd.

invoke_simd() passes pointers as generic address-space pointers,
and thus IGC BE has to generate runtime checks and conditional jumps
depending on which address space that pointer is (local vs global).

For a trivial esimd function called with uniform pointer for 1 block_load()
the code is 80 instructions instead of 69. Regression comes
not only from increase of number of instructions, but from conditions/jumps.

The fix is trivial. Because ESIMD works only with USM memory or with SLM,
when the block_load is not from SLM it must be the load from global memory
(address-space = 1). That address-space is forced in the fix in sycl-post-link
where the esimd intrinsics are lowered to vector load operations now.

Signed-off-by: Vyacheslav N Klochkov <vyacheslav.n.klochkov@intel.com>