[SYCL][NATIVECPU] Fix assert in MS name mangler (#11103)
This PR prevents the MS name mangler to assert on the unhandled
`CC_OpenCLKernel` calling convention which can end up in the AST when
targeting NativeCPU on Windows - see attached test case that triggers
the assert. This assert is also triggered on a number of e2e tests on
NativeCPU. This patch handles the CC_OpenCLKernel convention currently
only for NativeCPU and the usual error is raised for other targets as
before.