[SYCL][FPGA] Rename [[intel::disable_loop_pipelining]] attribute function metadata (#11372)
Metadata emitted from front-end when an FPGA IPA kernel has the property
pipelined<N> is a function metadata of the form !disable_loop_pipelining
!X, !X = !{i32 0|1}.

This is confusing as the user means to pipeline the kernel/not pipeline
the kernel as opposed to enabling/disabling the pipelining of all loops
in the kernel/function, and this name does not reflect that. This patch
renames the function metadata to "!pipeline_kernel", and to flip the
argument. We believe it may convey information better about what to do
with the kernel when the property is absent.

This was likely caused by a miscommunication with the related loop
metadata that's currently being attached when the loop attribute
[[intel::disable_loop_pipelining]] is used:
{!"llvm.loop.intel.pipelining.enable", i32 0}

---------

Signed-off-by: Soumi Manna <soumi.manna@intel.com>