[ESIMD][NFC] Remove compilation warning on FP atomics usage (#11530)
The FP atomic operations add the restrictions to the target device to be
at least ACM (any GPU that has LSC feature). This case is not the only
when using some features (types, or vector lengths) adds such extra HW
restrictions. There is no reason for fmax/fmin/etc to be special.

The warning used to signal about extra FP requirements was the obsolete
messages/warning, which is rather abuse of warnings system. Also, the
warnings printed for FP atomics were extremely annoying and polluted the
compiler output with no good benefit from it. As a result some customers
simply disabled 'obsolete' warnings completely, which is also bad as
they will not see really important warnings about API that is deprecated
and needs replacement in user's programs.

---------

Signed-off-by: Vyacheslav N Klochkov <vyacheslav.n.klochkov@intel.com>
Co-authored-by: Nick Sarnie <sarnex@users.noreply.github.com>