[SYCL][Graph] Reset L0 events for UR sync-points
The Unified Runtime sync-points used to represent
SYCL-Graph edges are not correctly implemented
for the Level Zero adapter. The L0 events used
to implement them are not reset to the non-signalled
state after an enqueue of the graph.

Fixed in the UR L0 adapter change https://github.com/oneapi-src/unified-runtime/pull/965
which resets the UR events to the non-signalled state at the
end of the L0 command-list.