[SYCL] Add support for -ftarget-register-alloc-mode (#11430)
We finally are using a new enough IGC driver in CI so we can implement
this. The option was approved a while back by the options group.

This change adds a new option `-ftarget-register-alloc-mode`. It is
specified as follows:

`-ftarget-register-alloc-mode=DeviceName0:Mode0[,DeviceName1:Mode1...]`

Currently the only valid device name is `pvc`, but I developed this
change to be extendable in the future.
The valid modes are:
auto -> request the backend to use heuristics to pick a register alloc
mode
small -> use small grf mode
large -> use large grf mode
default -> provide no specification to the backend on what register
alloc mode to use.

The default value if not specified is `pvc:default`, so provide no
specification to the backend in any case for any hardware. I will begin
an internal discussion after this is merged to see if we should change
the default to be `pvc:auto`.

The driver owns the mapping between the option mode and the backend
flag. It converts the user specified mode to the backend flag and passes
it either to ocloc for AOT or clang-offload-wrapper for JIT. The runtime
will extract the option stored in the image by clang-offload-wrapper and
pass it to the device backend if the device matches.

For AOT, we start making use of a new ocloc feature, `-device_options
<device> <options>`, ex `-device_options pvc "-options
-ze-opt-large-register-file"`.

This means "if the device is pvc, use large GRF mode. If the device is
not pvc, ignore the option."

For JIT, we do some argument processing in the runtime. The driver
splits the options in the case multiple are passed, and converts the
mode to the backend argument name to make the runtime processing easier.
For example, if the user passes
`-ftarget-target-alloc-mode=pvc:auto,pvc:large`, the options stored in
the image that the runtime sees will be
`-ftarget-target-alloc-mode=pvc:-ze-opt-large-register-file
-ftarget-target-alloc-mode=pvc:-ze-intel-enable-auto-large-GRF-mode`

The runtime loops over all the options, and for each instance of
`-ftarget-register-alloc-mode`:
Check if the device matches the device in the argument (currently only
pvc supported)
If so, replace the `-ftarget-register-alloc-mode=...` option with the
backend opt provided by the driver.
If not, remove the `-ftarget-register-alloc-mode=...` as to not pass it
to the backend.

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>