[SYCL][UR] Update CMake to enable easier local dev (#11492)
This patch introduces new ways to include the Unified Runtime source
code in the CMake build, the default behaviour is unchanged. It
introduces three new CMake variables:

* `SYCL_PI_UR_OVERRIDE_FETCH_CONTENT_REPO` is a variable which can be
used to override the `UNIFIED_RUNTIME_REPO` variable used by
`FetchContent` to attain the Unified Runtime source code.
* `SYCL_PI_UR_OVERRIDE_FETCH_CONTENT_TAG` is a variable which can be
used to override the `UNIFIED_RUNTIME_TAG` variable used by
`FetchContent` to attain the Unified Runtime source code.
* `SYCL_PI_UR_USE_FETCH_CONTENT` is an option to control if CMake should
use `FetchContent` to pull in the Unified Runtime repository, it
defaults to `ON`. When set to `OFF`, `FetchContent` will not be used,
instead:
* The path specified by variable `SYCL_PI_UR_SOURCE_DIR` will be used
with `add_directory()`. This can be used to point at an adjacent
directory containing a clone of the Unified Runtime repository.
* The path `sycl/plugins/unified_runtime/unified-runtime` will be used,
if it exists. This can be used as-if an in-tree build.
* `SYCL_PI_UR_SOURCE_DIR` is a variable used to specify the path to the
Unified Runtime repository when `SYCL_PI_UR_USE_FETCH_CONTENT` is set of
`OFF`.

Additionally:

* The logic using `UNIFIED_RUNTIME_LIBRARY` and
`UNIFIED_RUNTIME_INCLUDE_DIR` was removed because
`UNIFIED_RUNTIME_LIBRARY` has no other uses and the `if` statement would
always be true.
* The `unified_runtime` PI plugin is now always included in the CMake
build.