[SYCL][Graph] Finalize Graph for single device (#11565)
Only finalize a modifiable `command_graph` for the SYCL device that was
passed on construction. We
currently finalize for every device in the context, which is unnecessary
overhead and may lead to
unintended interactions with buffers used in these contexts.

PR also includes code to start waiting on for deps when adding commands
to a command graph. This prevents issues where allocation commands with
dependent copies could be delayed due to device being busy and execute
in an incorrect order with regards to future command graph executions.

---------

Co-authored-by: Ben Tracy <ben.tracy@codeplay.com>