[SYCL][Graph] Blocking wait in finalize on scheduler dependencies
When command-groups uses accessors in SYCL-Graph nodes, our implementation
uses the scheduler to wrangle dependencies and then add the node to the graph.
This happens at the application level during graph finalization. We currently
have a call to `waitForEvents()` when adding nodes, that waits on the device
event dependencies of that node command-group. For example, a memory copy command.

However, this corresponds to a call to `piEnqueueWaitForEvents`(), which is an
asynchronous call that itself returns an event that must be waited on later. The
blocking wait on this returned event is invoked by the scheduler on enqueue of
the executable graph.

This patch changes this behaviour to perform blocking waits earlier, on node
addition to the PI/UR command-buffer during graph finalization. We have designed
`finalize()` as the computationally expensive entry-point in the extension spec,
so blocking behaviour makes sense here rather than on executable graph submission.

There are two categories of blocking wait done when the scheduler adds a
node to the graph:

1. Blocking wait on host event dependencies, these correspond to scheduler
commands that don't have an associated PiEnqueue call. For example, memory
allocation commands in the scheduler. Adding this wait fixes the E2E failures
on DG2 devices, so the tests are re-enabled.

2. Blocking wait on device event dependencies, these correspond to scheduler
commands that do have an associated PiEnqueue call that returns an event. These
can occur when regular queue submissions are interleaved with adding graph nodes.
Introducing this wait fixes fails in the `buffer_ordering.cpp` E2E test on some
Level Zero devices.